requirements.txt
pyaudio>=0.2.14,<0.3
websockets>=14,<16
python-dotenv>=1.0,<2.0
requests>=2.32,<2.33
